requirements.txt
Flask==1.1.2
requests==2.26.0
jsonschema==3.2.0
scipy==1.4.1
imutils==0.5.4
opencv-python
dlib
Pillow
tqdm